move `data_governance` to top level in order to be reused
