+++ cyclonedx-bom/src/models/component.rs
@@ -590,14 +590,15 @@ mod test {
             attachment::Attachment,
             bom::BomReference,
             code::{Commit, Patch, PatchClassification},
+            data_governance::{DataGovernance, DataGovernanceResponsibleParty},
             external_reference::{ExternalReference, ExternalReferenceType, Uri},
             hash::{Hash, HashAlgorithm, HashValue},
             license::LicenseChoice,
             modelcard::{
                 ApproachType, ComponentData, ComponentDataType, ConfidenceInterval, Considerations,
-                DataContents, DataGovernance, DataGovernanceResponsibleParty, Dataset, Datasets,
-                Graphic, Graphics, Inputs, MLParameter, ModelParameters, ModelParametersApproach,
-                Outputs, PerformanceMetric, PerformanceMetrics, QuantitativeAnalysis,
+                DataContents, Dataset, Datasets, Graphic, Graphics, Inputs, MLParameter,
+                ModelParameters, ModelParametersApproach, Outputs, PerformanceMetric,
+                PerformanceMetrics, QuantitativeAnalysis,
             },
             organization::OrganizationalContact,
             property::Property,
