+++ cyclonedx-bom/src/specs/v1_5/modelcard.rs
@@ -27,13 +27,10 @@ use crate::{
     errors::XmlReadError,
     models::{self, bom::BomReference},
     prelude::Uri,
-    specs::common::{
-        organization::{OrganizationalContact, OrganizationalEntity},
-        property::Properties,
-    },
+    specs::{common::property::Properties, v1_5::data_governance::DataGovernance},
     utilities::{convert_optional, convert_vec},
     xml::{
-        optional_attribute, read_list_tag, read_simple_tag, to_xml_read_error, to_xml_write_error,
+        optional_attribute, read_simple_tag, to_xml_read_error, to_xml_write_error,
         unexpected_element_error, write_close_tag, write_simple_tag, write_start_tag, FromXml,
         ToInnerXml, ToXml,
     },
@@ -777,7 +774,7 @@ impl ToXml for ComponentData {
         }
 
         if let Some(governance) = &self.governance {
-            governance.write_xml_element(writer)?;
+            governance.write_xml_named_element(writer, GOVERNANCE_TAG)?;
         }
 
         write_close_tag(writer, DATASET_TAG)?;
@@ -1819,238 +1816,6 @@ impl FromXml for Graphic {
     }
 }
 
-#[derive(Debug, Deserialize, Serialize, PartialEq)]
-pub struct DataGovernance {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub(crate) custodians: Option<Vec<DataGovernanceResponsibleParty>>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub(crate) stewards: Option<Vec<DataGovernanceResponsibleParty>>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub(crate) owners: Option<Vec<DataGovernanceResponsibleParty>>,
-}
-
-impl From<models::modelcard::DataGovernance> for DataGovernance {
-    fn from(other: models::modelcard::DataGovernance) -> Self {
-        Self {
-            custodians: other.custodians.map(convert_vec),
-            stewards: other.stewards.map(convert_vec),
-            owners: other.owners.map(convert_vec),
-        }
-    }
-}
-
-impl From<DataGovernance> for models::modelcard::DataGovernance {
-    fn from(other: DataGovernance) -> Self {
-        Self {
-            custodians: other.custodians.map(convert_vec),
-            stewards: other.stewards.map(convert_vec),
-            owners: other.owners.map(convert_vec),
-        }
-    }
-}
-
-const CUSTODIANS_TAG: &str = "custodians";
-const CUSTODIAN_TAG: &str = "custodian";
-const STEWARDS_TAG: &str = "stewards";
-const STEWARD_TAG: &str = "steward";
-const OWNERS_TAG: &str = "owners";
-const OWNER_TAG: &str = "owner";
-
-impl ToXml for DataGovernance {
-    fn write_xml_element<W: std::io::Write>(
-        &self,
-        writer: &mut xml::EventWriter<W>,
-    ) -> Result<(), crate::errors::XmlWriteError> {
-        write_start_tag(writer, GOVERNANCE_TAG)?;
-
-        if let Some(owners) = &self.owners {
-            write_start_tag(writer, OWNERS_TAG)?;
-            for owner in owners {
-                write_start_tag(writer, OWNER_TAG)?;
-                owner.write_xml_element(writer)?;
-                write_close_tag(writer, OWNER_TAG)?;
-            }
-            write_close_tag(writer, OWNERS_TAG)?;
-        }
-
-        if let Some(custodians) = &self.custodians {
-            write_start_tag(writer, CUSTODIANS_TAG)?;
-            for custodian in custodians {
-                write_start_tag(writer, CUSTODIAN_TAG)?;
-                custodian.write_xml_element(writer)?;
-                write_close_tag(writer, CUSTODIAN_TAG)?;
-            }
-            write_close_tag(writer, CUSTODIANS_TAG)?;
-        }
-
-        if let Some(stewards) = &self.stewards {
-            write_start_tag(writer, STEWARDS_TAG)?;
-            for steward in stewards {
-                write_start_tag(writer, STEWARD_TAG)?;
-                steward.write_xml_element(writer)?;
-                write_close_tag(writer, STEWARD_TAG)?;
-            }
-            write_close_tag(writer, STEWARDS_TAG)?;
-        }
-
-        write_close_tag(writer, GOVERNANCE_TAG)?;
-
-        Ok(())
-    }
-}
-
-impl FromXml for DataGovernance {
-    fn read_xml_element<R: std::io::Read>(
-        event_reader: &mut xml::EventReader<R>,
-        element_name: &OwnedName,
-        _attributes: &[xml::attribute::OwnedAttribute],
-    ) -> Result<Self, XmlReadError>
-    where
-        Self: Sized,
-    {
-        let mut custodians: Option<Vec<DataGovernanceResponsibleParty>> = None;
-        let mut stewards: Option<Vec<DataGovernanceResponsibleParty>> = None;
-        let mut owners: Option<Vec<DataGovernanceResponsibleParty>> = None;
-        let mut got_end_tag = false;
-
-        while !got_end_tag {
-            let next_element = event_reader
-                .next()
-                .map_err(to_xml_read_error(&element_name.local_name))?;
-
-            match next_element {
-                reader::XmlEvent::StartElement { name, .. }
-                    if name.local_name == CUSTODIANS_TAG =>
-                {
-                    custodians = Some(read_list_tag(event_reader, &name, CUSTODIAN_TAG)?);
-                }
-
-                reader::XmlEvent::StartElement { name, .. } if name.local_name == STEWARDS_TAG => {
-                    stewards = Some(read_list_tag(event_reader, &name, STEWARD_TAG)?);
-                }
-
-                reader::XmlEvent::StartElement { name, .. } if name.local_name == OWNERS_TAG => {
-                    owners = Some(read_list_tag(event_reader, &name, OWNER_TAG)?);
-                }
-
-                reader::XmlEvent::EndElement { name } if &name == element_name => {
-                    got_end_tag = true;
-                }
-
-                _ => (),
-            }
-        }
-
-        Ok(Self {
-            custodians,
-            stewards,
-            owners,
-        })
-    }
-}
-
-#[derive(Debug, Deserialize, Serialize, PartialEq)]
-pub(crate) enum DataGovernanceResponsibleParty {
-    Organization(OrganizationalEntity),
-    Contact(OrganizationalContact),
-}
-
-impl From<models::modelcard::DataGovernanceResponsibleParty> for DataGovernanceResponsibleParty {
-    fn from(other: models::modelcard::DataGovernanceResponsibleParty) -> Self {
-        match other {
-            models::modelcard::DataGovernanceResponsibleParty::Organization(organization) => {
-                Self::Organization(organization.into())
-            }
-            models::modelcard::DataGovernanceResponsibleParty::Contact(contact) => {
-                Self::Contact(contact.into())
-            }
-        }
-    }
-}
-
-impl From<DataGovernanceResponsibleParty> for models::modelcard::DataGovernanceResponsibleParty {
-    fn from(other: DataGovernanceResponsibleParty) -> Self {
-        match other {
-            DataGovernanceResponsibleParty::Organization(organization) => {
-                Self::Organization(organization.into())
-            }
-            DataGovernanceResponsibleParty::Contact(contact) => Self::Contact(contact.into()),
-        }
-    }
-}
-
-const ORGANIZATION_TAG: &str = "organization";
-const CONTACT_TAG: &str = "contact";
-
-impl ToXml for DataGovernanceResponsibleParty {
-    fn write_xml_element<W: std::io::Write>(
-        &self,
-        writer: &mut xml::EventWriter<W>,
-    ) -> Result<(), crate::errors::XmlWriteError> {
-        match self {
-            DataGovernanceResponsibleParty::Organization(organization) => {
-                organization.write_xml_named_element(writer, ORGANIZATION_TAG)?;
-            }
-            DataGovernanceResponsibleParty::Contact(contact) => {
-                contact.write_xml_named_element(writer, CONTACT_TAG)?;
-            }
-        }
-
-        Ok(())
-    }
-}
-
-impl FromXml for DataGovernanceResponsibleParty {
-    fn read_xml_element<R: std::io::Read>(
-        event_reader: &mut xml::EventReader<R>,
-        element_name: &OwnedName,
-        _attributes: &[xml::attribute::OwnedAttribute],
-    ) -> Result<Self, XmlReadError>
-    where
-        Self: Sized,
-    {
-        let mut party: Option<DataGovernanceResponsibleParty> = None;
-        let mut got_end_tag = false;
-
-        while !got_end_tag {
-            let next_element = event_reader
-                .next()
-                .map_err(to_xml_read_error(&element_name.local_name))?;
-
-            match next_element {
-                reader::XmlEvent::StartElement {
-                    name, attributes, ..
-                } if name.local_name == ORGANIZATION_TAG => {
-                    let organization =
-                        OrganizationalEntity::read_xml_element(event_reader, &name, &attributes)?;
-                    party = Some(DataGovernanceResponsibleParty::Organization(organization));
-                }
-
-                reader::XmlEvent::StartElement {
-                    name, attributes, ..
-                } if name.local_name == CONTACT_TAG => {
-                    let contact =
-                        OrganizationalContact::read_xml_element(event_reader, &name, &attributes)?;
-                    party = Some(DataGovernanceResponsibleParty::Contact(contact));
-                }
-
-                reader::XmlEvent::EndElement { name } if &name == element_name => {
-                    got_end_tag = true;
-                }
-
-                unexpected => return Err(unexpected_element_error(element_name, unexpected)),
-            }
-        }
-
-        let party = party.ok_or_else(|| XmlReadError::RequiredDataMissing {
-            required_field: "organization or contact".to_string(),
-            element: element_name.local_name.to_string(),
-        })?;
-
-        Ok(party)
-    }
-}
-
 #[cfg(test)]
 pub(crate) mod test {
     use pretty_assertions::assert_eq;
@@ -2061,13 +1826,19 @@ pub(crate) mod test {
         specs::{
             common::organization::{OrganizationalContact, OrganizationalEntity},
             v1_5::modelcard::{
-                Attachment, Collection, ComponentData, ConfidenceInterval, DataContents,
-                DataGovernance, DataGovernanceResponsibleParty, Dataset, Datasets, Graphic,
-                Graphics, Inputs, MLParameter, ModelCard, ModelParameters, ModelParametersApproach,
-                Outputs, PerformanceMetric, PerformanceMetrics, QuantitativeAnalysis,
+                Attachment, Collection, ComponentData, ConfidenceInterval, DataContents, Dataset,
+                Datasets, Graphic, Graphics, Inputs, MLParameter, ModelCard, ModelParameters,
+                ModelParametersApproach, Outputs, PerformanceMetric, PerformanceMetrics,
+                QuantitativeAnalysis,
+            },
+            v1_5::{
+                data_governance::{DataGovernance, DataGovernanceResponsibleParty},
+                modelcard::GOVERNANCE_TAG,
             },
         },
-        xml::test::{read_element_from_string, write_element_to_string},
+        xml::test::{
+            read_element_from_string, write_element_to_string, write_named_element_to_string,
+        },
     };
 
     pub(crate) fn example_modelcard() -> ModelCard {
@@ -2149,12 +1920,12 @@ pub(crate) mod test {
         }
     }
 
-    pub(crate) fn corresponding_governance() -> models::modelcard::DataGovernance {
-        models::modelcard::DataGovernance {
+    pub(crate) fn corresponding_governance() -> models::data_governance::DataGovernance {
+        models::data_governance::DataGovernance {
             custodians: None,
             stewards: None,
             owners: Some(vec![
-                models::modelcard::DataGovernanceResponsibleParty::Contact(
+                models::data_governance::DataGovernanceResponsibleParty::Contact(
                     models::organization::OrganizationalContact {
                         bom_ref: Some(BomReference::new("contact-1")),
                         name: Some(NormalizedString::new("Contact")),
@@ -2236,7 +2007,7 @@ pub(crate) mod test {
 
     #[test]
     fn it_should_write_xml_data_governance() {
-        let xml_output = write_element_to_string(example_governance());
+        let xml_output = write_named_element_to_string(example_governance(), GOVERNANCE_TAG);
         insta::assert_snapshot!(xml_output);
     }
